% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HokMaxEnt-package.R
\docType{package}
\name{HokMaxEnt-package}
\alias{HokMaxEnt}
\alias{HokMaxEnt-package}
\title{HokMaxEnt: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
